pyproject.toml
[tool.poetry]
name = "neo-class"
version = "1"
description = "This project optimizes communication between teacher and student"
authors = ["Pavel Epanov <pavel.epanovm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.6"
fastapi = "^0.104.1"
minio = "^7.1.17"
sqlalchemy = "^2.0.23"
pydantic = "1.10.7"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
gunicorn = "^21.2.0"
python-jose = "^3.3.0"
redis = "^5.0.1"
fastapi-cache2 = "^0.2.1"
bcrypt = "^4.0.1"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
email-validator = "1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
